Added a bunch of functionality. The win screen currently isn't showing up due to weird positioning shit. Other then that everything is basically ok
End State: (currently shows you win when you shoot crystal but spacing is off so it might not appear in right place). I want it to transition to other screen, but this probably won't happen.

Multiple projectiles: Use w and q to switch between projectile types. Buttons remain the same.

Frozen Mode: When enemy is hit with ice it plays frozen animation. If frozen enemy is hit with ice a bonus is applied (currently +3 damage, but editable in inspector). With current levels, ice fire combo aloys kills enemy

Moving enemy: The small enemy now walks back in forth within a user defined area (currently 5) or turns when it hits anything. The enemy flips detection when it changes movement direction. Note: The actual area of movement is twice the allowed area entered in inspector. I could probably fix this but it seems like a small enough thing for designers to remember when designing.

Player facing direction of walking: The player turns to face in the direction that it is walking. However, the player can still shoot in a direction that they are not facing because the code to flip on shoot broke the flip on walk and made you always walk backwards. I'm sure this can be fixed with logic.